requirements.txt
apturl==0.5.2
asgiref==3.4.1
attrs==19.3.0
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.1
cachetools==4.2.2
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.6
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.7
django-allauth==0.45.0
django-cors-headers==3.10.0
django-environ==0.7.0
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
future==0.18.2
google-api-core==2.0.1
google-auth==2.1.0
google-cloud-core==2.0.0
google-cloud-secret-manager==2.7.1
google-cloud-storage==1.42.2
google-crc32c==1.2.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.40.0
httplib2==0.14.0
idna==3.2
importlib-metadata==1.5.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libcst==0.3.21
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
more-itertools==4.2.0
mypy-extensions==0.4.3
mysql-connector-python==8.0.26
Naked==0.1.31
netifaces==0.10.4
oauthlib==3.1.0
packaging==21.0
paramiko==2.6.0
pexpect==4.6.0
Pillow==8.3.2
proto-plus==1.19.0
protobuf==3.18.0
psutil==5.5.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycryptodome==3.12.0
pycryptodome-test-vectors==1.0.4
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python3-openid==3.2.0
pytz==2021.1
pyxdg==0.26
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
SecretStorage==2.3.1
shellescape==3.8.1
simplejson==3.16.0
six==1.16.0
sqlparse==0.4.2
ssh-import-id==5.10
system-service==0.3
systemd-python==234
texttable==1.6.2
typing-extensions==3.10.0.2
typing-inspect==0.7.1
ubuntu-advantage-tools==27.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.6
usb-creator==0.3.7
wadllib==1.3.3
websocket-client==0.53.0
xkit==0.0.0
zipp==1.0.0
